+++ lib/medic/sugar.ex
@@ -1,0 +1,5 @@
+defmodule Medic.Sugar do
+  @moduledoc false
+
+  def ok(term), do: {:ok, term}
+end